Deduplicate Makefile writer and Jinja parameter filter
The Makefile-writing logic existed twice, byte for byte: in
EdaCommands.write (used by legacy backends) and in the make build
runner (used by the flow API). A fix in one could not reach the
other. Make.write now delegates to EdaCommands.write.

jinja_filter_param_value_str was likewise defined identically in
edalize/edatool.py and edalize/tools/edatool.py. Move the single
definition to edalize/utils.py and import it from both.

No functional change; golden-file tests confirm generated output
is identical.
